requirements.txt
mysqlclient
python-decouple==3.4
